docs: complete milestone widget-cleanup
- All 4 target files now ≤250 lines
- Extracted TemplateCountBadge + CategoryChips from home_screen.dart (270 → 160)
- Trimmed register_screen.dart (253 → 248)
- 606/606 tests passing, analyzer clean
- Milestone summary archived to .gsd/milestones/
