+++ .gsd/STATE.md
@@ -1,19 +1,12 @@
 # STATE.md
 
 ## Current Position
-- **Milestone**: Widget Cleanup
-- **Phase**: 1 — in progress
-- **Status**: 4/4 file targets met, awaiting test & analyzer verification
+- **Milestone**: None active
+- **Last Completed**: Widget Cleanup ✅ (2026-02-19)
 
-## Phases
-
-| Phase | Name | Plans | Status |
-|-------|------|-------|--------|
-| 1 | Theme & Screen Extraction | — | 🔄 In Progress |
-
-## Previous Milestone
-Codebase Improvement ✅ — completed 2026-02-19 (tag: `codebase-improvement`)
+## Previous Milestones
+- Widget Cleanup ✅ — completed 2026-02-19 (tag: `widget-cleanup`)
+- Codebase Improvement ✅ — completed 2026-02-19 (tag: `codebase-improvement`)
 
 ## Next Steps
-- Run full `flutter analyze` and test suite to close remaining must-haves
-- Mark Phase 1 complete once all checks pass
+- Pick next milestone from backlog
