+++ .gsd/milestones/widget-cleanup-SUMMARY.md
@@ -1,0 +1,33 @@
+# Milestone: Widget Cleanup
+
+## Completed: 2026-02-19
+
+## Deliverables
+- ✅ `app_component_themes.dart` ≤250 lines (302 → 228)
+- ✅ `home_screen.dart` ≤250 lines (270 → 160)
+- ✅ `create_screen.dart` ≤250 lines (270 → 246)
+- ✅ `register_screen.dart` ≤250 lines (253 → 248)
+- ✅ `flutter analyze` clean on all modified files
+- ✅ All 606 tests passing
+
+## Phases Completed
+1. Phase 1: Theme & Screen Extraction — 2026-02-19
+
+## Extracted Components
+| Component | Source | Destination |
+|-----------|--------|-------------|
+| `AppButtonThemes` | `app_component_themes.dart` | `app_button_themes.dart` |
+| `GenerationStartingOverlay` | `create_screen.dart` | `generation_starting_overlay.dart` |
+| `TemplateCountBadge` | `home_screen.dart` | `home_screen_widgets.dart` |
+| `CategoryChips` | `home_screen.dart` | `home_screen_widgets.dart` |
+
+## Metrics
+- Total commits: 2 (PR #17 squash merge + final fixes)
+- Files changed: 8
+- New files created: 3
+- Duration: 1 day
+
+## Lessons Learned
+- Some files appeared compliant on feature branch but not on master — always verify on master after merge.
+- Pure structural refactors (copy-paste extraction) are low-risk but still need import/export verification.
+- Re-export pattern (`export 'file.dart'`) prevents consumer breakage when splitting files.
